Add Tiny Swords asset pack
By downloading and unzipping latest Tiny Swords (Update 010) from
Itch.io [0]. This pack has Creative Commons Zero (CC0) license so no
attribution is required in this commit.

Ignore Aseprite working files and macOS specific files.

This is intended to be a base ground or placeholder art, but sets the
project format to:

- Tilemap grid: 64x64 pixels.
- Animations speed: 10fps or 100ms.

Importing all the PNG textures into the Godot project for simplicity.
They should be limited to the ones appearing in the game before
releasing.

[0] https://pixelfrog-assets.itch.io/tiny-swords