Add talker NPC
Talker class that inherits from NPC. In turn, NPC inherits from
CharacterBody2D. This is to make room for enemies, which are going to be
NPCs as well.

Talker has a name and a dialogue to be started upon player interaction.

NPC has a "look at" side for turning left or right. And a sprite frames
for changing the sprite asset. Two sprite frames resouces are provided
using the Tiny Swords pack.

Make these tools, so look at and sprite change are reflected in the
editor.

Use collision layer 2 for NPCs, rename it accordingly. And make it
collide with the walls and the player.

Add 2 instances of the NPC to the main scene.

Add a new Node2D called "OnTheGround" in the main scene to contain the
player and the NPCs, and potentially all elements that need to be on the
ground. And enable Y sort ordering so the elements for a top-down
effect.