Add Player / NPC interaction
Use a RayCast2D and an Area2D to detect if a game entity can interact
with another. The interaction may not be immediate, (eg. when they
trigger a dialogue). Add signals so the both entities can check when the
interaction has started and when it has ended.

Add an interaction ray to the Player, and an interaction area to the
talker NPC. Trigger the talker dialogue when interacting. The player
skips movement during the dialogue.

Reorder player scene and scripts into its own folder. Split the scripts
into separate components. Otherwise the Player _process was doing too
much.