Add musical rocks puzzle
This adds a series of rocks which the player can interact with to play
musical notes. These are used for a puzzle where, if the player kicks
the rocks in a certain series of sequences, they complete a puzzle and
gain a reward. The player can interact with an NPC who tells some
backstory, demonstrates the first part of the melody, and directs the
player to some environmental clues for how to form the remaining
melodies.

The concept is inspired by puzzles in Stardew Valley and Return to
Monkey Island, with the environmental clues lovingly stolen from the
Crystal Heart puzzle in chapter 6 of Celeste.

The components of the puzzle are structured with the general intent that
they could be reused by other puzzles:

- The note played by each rock (both its audio sample and its note name
  (A-G)) are properties of the rock, with the colour of the rock being
  derived from the note name.

- The xylophone iterates its children to find all available notes, so
  notes can be added and removed by adjusting the rock child nodes.

- The tunes to be played are a property of the MusicPuzzle scene, which
  tracks the player's input, ignites fires (themselves another property)
  corresponding to each step of the puzzle, and emits a signal when the
  puzzle as a whole is solved.

The NPC -> Xylophone interaction is a little ad-hoc (using mood tags)
but does work as one approach to having dialogue which interacts with
the outside world.

The reward for completing the puzzle is that a bouncing sheep appears.
(Maybe the player can shear it to make wool.) Of course, like the
dialogue, assets, etc., this is a placeholder!

https://github.com/endlessm/threadbare/issues/5