Make music puzzle more forgiving
Suppose you have just solved the first melody and are currently trying to
play DEFA to solve the second melody.

Previously, if you played DDEFA, the game would not accept your solution:

- The initial D was accepted as the first note of the expected melody.
- The second D was matched against the second note of the expected
  melody, E. It did not match, resetting the position to the start of
  the melody.
- The E was matched against the first note of the expected melody, D,
  and rejected. Ditto the F and A.

To fix this, when resetting the position, match the given note once more
against the start of the pattern. With this change, the input DDEFA is
accepted:

- The initial D is accepted as the first note of the expected melody.
- The second D is matched against the second note of the expected
  melody, E. It does not match, resetting the position to the start of
  the melody; but then, it is matched against the first note of the
  expected melody, D, and accepted.
- E, F, A are then accepted, completing the puzzle.