Player: Enable hit_from_inside on interact ray
Previously, if the player is positioned such that its InteractRay starts
inside an NPC's InteractArea, the ray would not be considered to have
collided with the NPC so the interaction would not be offered. This can
happen because we are faking perspective by giving NPCs (and the player)
two hitboxes: one for collisions around the feet, and one for
interactions in the player's line of sight.

Enable RayCast2D.hit_from_inside on the player's interact ray.