Add splash screen with animated Endless logo
This is meant to simulate, with my limited artistic skill, an invisible
person sewing the Endless logo into the page with a sewing machine,
drawing it with a series of straight lines that are meant to look like
stitches.

The path and width curve are based on me tracing the logo by hand. It's
not precise but that is OK: it's meant to look human-made, and there is
randomness added to the endpoints of each stitch.

The duration of the animation is synced to the length of a sound effect
of a sewing machine.

The "real" logo is included by way of reference. Details of the source
of other assets are in the respective README files.