Join world map, rocks puzzle & stealth challenge together
Add a subclass of Talker with a method ("mutation" in Dialogue Manager
speak) which, when invoked, switches to a new scene when the dialogue
ends. Add a dialogue template for this quest-starter NPC which uses the
NPC's name and a "quest description" property to introduce the quest,
the idea being that learners need not learn how to edit dialogue to add
a way into their quest.

Add an NPC to the Song Sanctuary using this new script that transports
the player to the musical-rocks puzzle.

Add a teleporter in the same location as the placeholder bouncing-sheep
which takes the player to the stealth quest.

Change the NPC at the end of the stealth quest into a quest-starter,
transporting the player back to the Song Sanctuary.

Now that you can enter a quest from the world map, the game can place
the player into the world map after the introductory sequence.

Fixes:  https://github.com/endlessm/threadbare/issues/32