music_puzzle: Simplify the bard playing the xylophone
Previously, the communication between the bard dialogue and the
xylophone was managed by using emotion tags in the dialogue, which were
then detected by the bard's script and transformed into xylophone actions.
This was a bit convoluted and also made it impossible to synchronise the
notes with the dialogue text.

Now that the Talker node (in this case the Bard) is passed to Dialogue
Manager as extra game state, the dialogue can call methods on the bard.
As a bonus, if the method is a coroutine (i.e. a function that uses
`await`), typing of dialogue will pause until the coroutine resolves.

Define a play() method on the bard, which is not necessary but makes the
dialogue more readable. Call it from the dialogue in place of using
tags. Remove the special handling of dialogue tags.

https://github.com/nathanhoad/godot_dialogue_manager/blob/main/docs/Conditions_Mutations.md#mutations