music_puzzle: Adjust dialogue based on puzzle progress
Previously, the bard would always have the same conversation with the player.

With this change, the bard's dialogue varies depending on a number of conditions:

- If the player has spoken to the bard before, the dialogue is shortened
  this lets the player see the melody again without having to go through
  the whole conversation.

- If this is the first interaction, but the player has already guessed
  the first line of the melody, the bard will take notice.

- If the player has solved the puzzle, the bard will congratulate them.

This involves adding a boolean to the bard script. This may not scale;
in the general case it may be useful to add a dictionary of dialogue
state to each NPC, or save this state more globally.