Simple global state to store items and future globals
After some back and forth with @JuanFdS, we came up with an approach that we believe is user friendly for the Learners, if they need to use global state.

We have a global scene called GameState which, at the moment, only has an "inventory", which in turn has "items". We thought that this results in code-as-documentation, where you can know what kind of global state there's by looking into the code of game_state.gd.

Items can be created as resources and they can be, for the time being, of any of the three types: Memory, Imagination, or Spirit.

The global scene can be edited using godot inspector, which can be useful to test different states.
![image](https://github.com/user-attachments/assets/4a0db1db-a7ea-4511-b991-c760bd381afe)

As for persistance, there's none at the moment (other than editing the scene manually, which is fixed at compiled time). But in the future, a serializer / deserializer could be implemented that sets the global state based on some conventions, to ensure that learners can add global state without much hazzle and have that be persisted.