intro: Replace placeholder text with dialogue
Previously, a one-line placeholder introduction to the world was
displayed by animating the visible_ratio property of a RichTextLabel.

Instead, define the on-screen introduction text as a dialogue file.
Trigger the fade-out animation with a mutation (a.k.a. method call) in
the dialogue file. Replace the RichTextLabel with an invocation of that
dialogue.

https://github.com/endlessm/threadbare/issues/6