Added helper script for PointLight2D to make it easier to visualize and get colliders for it
The Guards detection area should match the light they cast so the
players can understand where they will be detected by guards and where
they would be safe.

PointLight2D uses a texture for the shape of the light, and the light's
shape isn't easy to see unless there's a there's a visible node
that's being lit by that light. In order to make the light's shape
easier to modify in the editor, added a debug toggle that draws a Sprite
with the same texture of the light.
Also, added a button that creates a collision based on the light's
texture, so if the texture is changed, the collision can be changed
accordingly without too much work.