Player: Animate sprite through AnimationPlayer
Add a new AnimationPlayer node to the Player scene, and do the current
idle and walk animations using it. This is because the upcoming attack
must have the frame-by-frame animation in sync with collision shapes
animations. And overall is better to have the frame by frame animation
in a timeline to add sound FXs, etc.

Also, don't make character configurable in player sprite. Realistically
the player sprite is only for the player and will soon have
player-specific animations. So add a class name to Player and
remove the export from PlayerSprite. Assume that the owner is the player
when the node is ready.