Restore and save current scene to URL hash in web builds
During development, it is very useful to be able to link directly to a
scene, so that a pull request can be easily playtested.

To enable this, extend SceneSwitcher to update the URL hash when
switching scenes, if running on the web platform. On startup, check the
URL hash, and if it matches a scene, switch to it. Later changes to the
URL hash are ignored.

The scene can be specified in uid:// or res:// form (or anything else
that ResourceLoader understands). When changing scene (including to the
one specified in the URL hash), the hash is set based on the
PackedScene.resource_path property, which is the more human-readable
res:// path form.

If the scene path begins with res://scenes/ and ends with .tscn, this
prefix and suffix are stripped for a more human-readable URL. On
startup, if the URL hash looks like a relative path (i.e. does not begin
with res:// or uid://), these are readded.

This is based on code I originally wrote for Candy Collective:
https://github.com/endlessm/candy-collective/blob/main/Script/SceneSwitcher.gd

Fixes https://github.com/endlessm/threadbare/issues/40