Player: Add fighting player
Duplicate the player for now to add ink fighting features.

Add a health bar scene, which is just a ProgressBar node with custom
style. This will be reused in enemies too.

Add an instance of that health bar scene to the player. It will be
invisible until the player is engaged in a fight (when the
player gets the first hit).

Add a HitBox Area2D to the player, to receive hits from enemies.

Add player health variable and reduce it when hit. Use a constant harm
value for now.

Restart the current scene when the player health reaches zero.