intro: Remove [next=auto] directives
I had included [next=auto] directives at the end of each line of the
introduction so that the intro text would advance automatically after a
few moments (proportional to the length of the text).

What I had not appreciated is that if you configure the text to advance
automatically, the player can't press ui_accept (space/A) to move
forward more quickly. This is annoying if you want to skip the
introduction.

Remove these directives.