Player: calculate interact area when requested by player_interaction
Needed to make this change because I wanted to free a node that parented an interact_area and ran into an error. The way to reproduce that error was:
- Have an interact area in range of player's interact ray.
- Queue free the interact area (in my case, this happened as a result of the interaction).

If I understood correctly, the reason for this error was that depending on the order in which the _process and the freeing of the node were executed, the interact_ray could keep a reference to an already deleted node and that was accesed by player_interaction before interact_ray could reassign that variable to null.