Ink combat: Add ink follow player feature
The ink can now have a Node2D to follow. A constant force in that node
direction will be set every frame.

The enemy has a boolean export to throw ink that follows the player node
(the first and only node that should be in the "player" group).

Fix https://github.com/endlessm/threadbare/issues/71