Added Pause System & Pause global
The Pause System allows pausing the whole game, or just the player input, and is flexible enough to allow adding other kinds of pauses. It also supports being requested pauses from a multitude of different origins, and handling it gracefully. This pattern of requests prevents some bugs that could appear if the pause logic becomes more complex. For instance let's say we have an inventory and a pause menu. And both pause the game. The pattern ensures that as long as one is opened, the game is paused. Another safety measure: Each pause request is attached to a node. If that node leaves the tree, then it is assumed that the request should be cleared.

This commit also adds a global singleton called Pause, which handles the pause requests and pausing / unpausing. This has to be global because pauses can affect scenes right after they load (for instance, during a transition), and also, because the mechanism for pausing the whole tree in Godot is global and trascends the current scene.