Remove date from first-party FileCopyrightText notices
Previously, the codebase was inconsistent. Some files used:

    SPDX-FileCopyrightText: 2025 The Threadbare Authors

while others used:

    SPDX-FileCopyrightText: The Threadbare Authors

As argued in https://daniel.haxx.se/blog/2023/01/08/copyright-without-years/
there is no strict need to include the year in copyright notices – the
information is present in the Git commit history. And excluding the year
means that there is no need to think about bumping the upper bound of
the range each time a file is modified in a new year.

Remove the year from first-party copyright notices.