Make interact label position work even if camera's zoomed
ScreenOverlay had follow_viewport enabled set to true, so it is already passing a transform to its children. Changed the player interaction code so it doesn't try to compensate for the canvas layer, since ScreenOverlay is already handling that with the follow_viewport setting.

Also, replaced every instance of the ScreenOverlay canvas layer by a scene, so if we need to change any setting is done just once.