Player fighting: Remove first anticipation
So the player character blows ink immediately when the action is
pressed. Then the anticipation happens in the successive attacks, when
the action keeps being pressed. The anticipation in the successive
attacks can be skipped by smashing the button, but that should be OK for
now.

Also move the anticipation time to a constant.