Ink combat: Add ink blob hit sound effect
When the ink blob is hit by the player so it can hit enemies and
inkwells, the ink blob changes its animation. But there wasn't an
audible clue. Add one.

Also, reuse this sound when the ink blob appears, so the enemies
spawning ink blobs also has an audible clue.