Ink combat: Remove enemies after game complete
Removing enemies after completing an inkwell dates back to the goal of
clearing the enemies. Now the goal is to fill the inkwells, and the game
becomes easier when they are removed one by one.

Do not remove a random enemy each time an inkwell is completed. Instead,
make them dissolve after the goal is reached. That is, after a certain
amount of inkwells have been finished. So the player can reach the
collectible.