Reorganize project files
Based on what was discussed in https://github.com/endlessm/threadbare/issues/75, the changes are:

In scenes/, have the folders

 - story_quests, where we will store the playable levels that make up each story quest.

 - world_map, here will live the scenes that make up the open world of the game.

 - game_elements, reusable elements (enemies, npcs, props) that are the building blocks of any playable scene.

 - menues (pause, main_screen, credits, etc) are the scenes that aren't playable and also are mainly UI centric.

 - ui_elements, the building blocks of the menues scenes, and also extradiegetic elements like huds.

 - globals, nodes that live outside of any scene but are needed for the game to work.

Also, another decision taken was to have the .tscn files apart from the other files that make up a scene (script files, assets, resources, etc). The idea behind this is that we want learners to be able to quickly figure out what's the 'ready to use' file that they can drag to the editor.