SceneSwitcher: Add change_to_packed_with_transition
Previously, this module exposed a function to change to a PackedScene
without a transition, or to change to a scene by path with a transition.

Add a new variation that changes to a PackedScene with a transition.
Refactor the existing change_to_file_with_transition() function to call
the new variation.