Make it possible to remove items from the inventory.
The items are also removed from the HUD if the removed items where the same.

This can also handle an inventory where there were items from multiple story quests.

As long as an item of the same type as the one in the progress HUD is removed, the progress HUD slot will be emptied.