Player: Add template assets
Add template spritesheet PNGs for the animations: idle, walk, attack_01,
attack_02.

These are supposed to be basic drawings that a learner can color and
draw on top.

Add a SpriteFrames resource that uses these PNGs.

Compared with the Story Weaver, the frames of some animations have been
reduced, to reduce the modding work:

- idle: from 10 to 3
- walk: from 6 to 3

Still, the SpriteFrames has the correct number of frames, because some
frames are duplicated.

Add a thin 1-pixel line between frames in the sheets, to visually show
where frames start and end when editing the PNGs. Because of this, the
frames have to be imported like this in the SpriteFrames editor:

- Size: 192 px by 192 px
- Separation: 1 px