Ink combat: Make Enemy throw ink of matching color
The enemy picks a random color of an array of allowed colors. The
initial array has all colors. But the combat logic checks which is the
ink needed to complete the goal, and updates the enemies to throw the
wanted colors.

This improves the combat because the player doesn't have to wait for the
needed color. In a easy combat, plenty of barrels of the same color can
be placed in the scene. Then the enemy will throw only that color.