Splash: Switch to intro with fade transition
Previously, when the splash screen was complete (either due to 5 seconds
passing from the end of the animation, or the player pressing a key),
the scene would be switched abruptly to the intro, without any
transition.

Transition with a fade - this seems more appropriate than a wipe because
the two screens are not spatially connected.

The switch can be triggered in two ways:

1. The animation completing, and then 5 seconds elapsing;
2. The user pressing a button.

If the user presses a button 4½ seconds after the animation completes,
we do not want to perform the scene switch twice. To deal with this,
move the timer into the scene; and disconnect from its "timeout" signal
when starting a transition for any reason.