Add a player direction to spawn point
This commit adds a player direction to the spawn point. When set to other than "UNSPECIFIED", it will change the player look at direction to either left or right.

When players teleport from one scene to the other, sometimes it looks weird if the player looks in the opposite direction. This solves that issue by allowing to specify the look direction in the spawn point.

Extracted LookAtSide enum to a global so it can be reused throughout the project.