Add closing transition scene for lore quest
Create a new scene that is purely narrative and has no gameplay.

This scene contains the dialogue proposed in this document: https://docs.google.com/document/d/1aHvhG8WJ5zahHbLdf_5WD7uYrwLIUGes1GtifTp04aM/edit?tab=t.0#heading=h.rlrm1ccl0zl5

There is a little cinematic happening in this scene, which was achieved by using the sprite of the player without the player itself being on scene. This could fall short later if the sprite isn't enough to represent the StoryWeaver in the future (for example, if we add dust particles on walk, that won't be here), but I opted for making it simpler, so this scene can showcase how a simple cinematic can be achieved with AnimationPlayer and some scripting, which could be useful for learners.