Intro: Refactor into reusable Cinematic script
Currently, we have two very similar implementations of a cutscene in the
project: the current intro scene, and the closing transition scene of
the MVP quest. Both of these play a simple dialogue, synchronised with
some animations in an AnimationPlayer, then transition to another fixed
scene.

In the near future, there will be more such cutscenes: one at the start
and end of each StoryQuest, each of which will take place in a separate
narrative universe so needs its own introduction & conclusion.

Refactor the current intro.gd script into a reusable cinematic.gd
script, with three exported properties:

- The dialogue to play (with a placeholder dialogue by default);
- The AnimationPlayer node, if needed for the dialogue;
- The scene to switch to at the end of the dialogue.

As with the current implementation, the cinematic node itself is exposed
to the dialogue.

Restructure the current Intro scene tree, which was previously a Control
with two children, to have a Node2D at the top-level, and another Node2D
holding the Cinematic script as a child, mirroring the MVP quest's
closing transition.

Call animation_player.play() directly from the intro dialogue, rather
than providing a method on the intro scene.

Simplifying awaiting the end of the dialogue with `await`: this is how
closing_transition_cinematic.gd implements the same thing and it's
neater.

Move the intro's 1-second delay to AnimationPlayer. This delay was added
in commit 1f4ffc918a4cf537ed39bd8d2316da1c1df1584c to not start playing
dialogue until the scene-switching fade transition is complete.

closing_transition_cinematic implements very similar logic, except that
it does so by playing a "walk into the scene" animation that can
naturally occur while the fade-in is playing. Do the same here, with a
1-second no-op animation that is triggered from, and awaited by, the
dialogue. No balloon is shown until the first actual line of dialogue,
so this has the same effect as the previous logic.