Dialogue: Invert text colour in buttons
Prevously, the default font colour was used, which is either white or
slightly off-white depending on the state of the button. This is not
legible against the background colour of the button assets.

Invert each colour. For instance, Button/colors/font_color defaults to
hsv(0, 0, 88), so here it is changed to hsv(0, 0, 100 - 88) = hsv(0, 0,
12).

https://github.com/endlessm/threadbare/issues/205#issuecomment-2809899852