Add methods to property utils to expose children properties
In order to give level editors the ability to edit some parts of a particular scene, without having to use "editable_children", sometimes as a developers you want to expose properties of child nodes as if they were of the parent.

By using the methods added in this commit in "_get_property_list()" it becomes trivial to expose children properties with little boilerplate.