Guard: modify creation of new patrol path to ensure it has 2 points
Previously, if a guard didn't have a patrol_path and the Add/Edit Patrol Path button was used, it would create a new path with an empty curve. Now, it creates a new path with a curve that has 2 points, one at the position the guard is at, and the other point at a random position that is at 150px from the guard.