Guard: highlight patrol path when a guard is selected in the editor
It wasn't clear which paths belonged to which guards by just looking at the layout of a level. Now, when a guard is selected in the scene tree, a red line will be drawn over its patrol path (if it has one), and red circles will be drawn at the points of that path where the guard stops.

Related to https://github.com/endlessm/threadbare/issues/228