Dialogue: Add visible “next” Button
In playtesting we have found that not all players realise that they need
to press a key to progress the dialogue.

Add a button that, when pressed, acts like the input action for "next"
(Enter, Space, Xbox joypad A) has been entered.

Style variations for the button added in a new style called FlatButton.

Fixes https://github.com/endlessm/threadbare/issues/263