tileset: Add a cliff-face alternative without collision
This is an alternative tile to the non-edge cliff front. It's needed
when a bridge reaches land from the south, above a cliff face;
otherwise, using the normal impassable cliff-front tile, it would be
impossible to have the cliff face below the bridge and still allow the
player to walk across it.