Add pre-commit exception for script_templates/
I think it's not necessarily an issue if a script template doesn't pass gdlint since it will have to be modified and completed (so checking unused arguments at that point won't always make sense) and also because script templates aren't valid gd files. Finally, I don't know if there's a way to have the meta-description (which is set as a comment) have multiple lines, so the max allowed line length check will also cause troubles.