Add property to interact_area.gd that indicates where the interaction label should be placed
Instead of drawing the interaction label over the player, it is drawn at an offset of the interact_area called interact_label_position.

Added code in interact_area to display approximately where the interact_label would appear when a player is near.

Why use a Vector2D property instead of a child node (like a Marker2D)? Because we are adding InteractAreas in several places by creating nodes from the script instead of dragging the interact_area.tscn, and I didn't think it was worth it to break that for this change.

Related to https://github.com/endlessm/threadbare/issues/167